Separate web and lib compilation (#24)
We create a new build process for web that uses `npm`, which is more widely supported for building for the web. With the old, Deno-based, system, for example, using recent versions of Sentry was causing the build to explode. This requires us to use `dnt` to compile our libs in a Deno-compatible way, and then use esbuild on the web `main.ts`.